+++ src/app/service/vendor-service.ts
@@ -1,8 +1,9 @@
 
 import { Injectable } from "@angular/core";
 import { HttpClient } from "@angular/common/http";
-import { Observable } from "rxjs";
+import { Observable, throwError } from "rxjs";
 import { Vendor } from "../model/vendor";
+import { AuthService } from './auth-service';
 
 const URL = "http://localhost:8080/api/vendors";
 
@@ -11,17 +12,28 @@ const URL = "http://localhost:8080/api/vendors";
 })
 export class VendorService {
 
-  constructor(private http: HttpClient) {}  
+  constructor(private http: HttpClient, private authService: AuthService) {}  
 
   list(): Observable<Vendor[]> {
+    // Regular users can't view the full vendor list
+    const user = this.authService.getCurrentUser();
+    if (!user?.admin) {
+      return throwError(() => new Error('Insufficient permissions to view vendors'));
+    }
     return this.http.get<Vendor[]>(URL + '/');
   }
 
   add(vendor: Vendor): Observable<Vendor> {
+    if (!this.authService.hasPermission('addVendor')) {
+      return throwError(() => new Error('Insufficient permissions to add vendor'));
+    }
     return this.http.post<Vendor>(URL, vendor);  
   }
 
   update(vendor: Vendor): Observable<Vendor> {
+    if (!this.authService.hasPermission('updateVendor')) {
+      return throwError(() => new Error('Insufficient permissions to update vendor'));
+    }
     return this.http.put<Vendor>(URL + '/' + vendor.id, vendor); 
   }
 
@@ -30,6 +42,9 @@ export class VendorService {
   }
 
   delete(id: number): Observable<Vendor> {
+    if (!this.authService.hasPermission('deleteVendor')) {
+      return throwError(() => new Error('Insufficient permissions to delete vendor'));
+    }
     return this.http.delete<Vendor>(URL + '/' + id);
   }
 }
